Add autosuggestions in chat bar
This will add support for showing autosuggestions. If text input contains a prefix and if that prefix is supported
then the suggestions will be shown. And as user types we'll update the suggestions by showing only those suggestions
whose key contains user entered text.

We are creating a tableview in the conversation VC and passing it to the chat bar as chat bar doesn't know where these
suggestions will be displayed. Tableview's datasource and delegate is implemented in the chat bar as text is handled
in the chat bar. Chat bar will get the data(autosuggestions) and prefix from outside.